go back to intLang with optional labels for bodies
Aim is to keep the proofs up to exp_to_Cexp intact, since labelising can
be done separately and it's not worth breaking the proofs to do it at
the same time as exp_to_Cexp.

This is based on code in 9fedec64b61ad3e35a848b4ed6d769dac00401a2 in the
HOL repository. (This repository's history doesn't go back far enough to
contain the old intLag.)

Working on a new, simpler approach to labelising:
Pull any unlabeled bodies into the code env (but without
recursively doing the same to the bodies).
Repeat on each exp in the range of the code env (merge the resulting
code env in), until no unlabeled bodies remain.

In this commit:
Old correctness proofs (up to exp_to_Cexp) work.
Defined new labelise, but haven't proved anything about it (including
termination) and compiler tests are broken.