Generalise subst_labs_any_env
Changed the definition slightly too, so it leaves a label in place if it
doens't occur in the code environment's domain.

Also, the start of an attempt to extend label_closures_thm to also
characterise returned exp (in terms of subst_labs).

Also start of attempt to say something about Cevaluate_subst_labs.