finish CCon case of compile_val
We now need to track a good state down all the sub-expressions as they
are evaluated, and then get a further good state when their values are
combined, because the store might change at every step.

Previously, we just used the good state at the beginning as a basis to
conclude that the final state (with one extra thing on the stack) is
good. That shortcut doesn't work any more, unfortunately.