finish CLet case of compile_val
As it turns out, propagating a good Cenv_bs correspondence down each
step of evaluation is not as necessary as the last commit message would
suggest, since it is possible to change the store argument of Cenv_bs
arbitrarily as long as you can prove that the store correspondence
(s_refs) holds. See the new theorem Cenv_bs_change_store, a trivial
consequence of Cenv_bs_def.

Propagating the invariant properly would be very difficult in the CLet
case, so this is lucky.