thoughts about compile_val opapp case, define good_code_env
good_code_env says every expression in the code environment has the
property that if it evaluates to a value in an environment that came
from a closure, and the machine looks like its been set up for a call,
then calling the pointer to that expression will eventually give the
same value.

right now, trying to get good_code_env right to be used as an
assumption. but the real difficulty will be when it comes to having to
prove it (probably only when creating closures).