attempt to fix compile_val statement to allow refs to change
The old statement had been provable because reference-related
expressions are still not in the language subset.

This attempt has revealed a relatively deep problem: in a stateful
language, I reckon I can't use a previous good state as proof that a
final state is good without going through all the intermediate states.
For Let expressions, then, shadowing causes a problem because the
shadowed variable disappears from the goodness invariant in some of the
intermediate states, and I don't know how to get it back when the Let
closes.

Assuming there's no way to solve that, there are at least two ways to
get rid of shadowing:
  1. make all variable names in the intermediate language distinct
  2. use an environment representation that keeps all bindings around
I'm more inclined to do 1, which could be via deBruijn indices, or just
renaming variables on conversion into the intermediate language.