prove CLet case by assuming distinct bound variables
The hope is that we can discharge these extra assumptions with a pass
over the intermediate language that renames binders. Indeed it might be
sufficient to say that such a pass is always possible, without actually
doing it.

But first, should find out if the rest of compile_val goes through with
these extra assumptions about distinct binders.