more cases of compile_val, stuck on list case
Because Cv_bv with one refs does not necessarily imply Cv_bv with
another refs. May be able to fix this by changing the statement of
compile_val in the list case: require a corresponding machine state
after every expression, rather than just at the end of the list of
expressions.