changing refs should not affect Cv_bv; start work from this insight
the only reason Cv_bv has access to the refs is so it can verify that
RefPtrs in closures point to the correct things. These RefPtrs are not
touchable by the source language, so as long as we can prove that they
don't change (although more can be added, when closures are added!)
should be able to change the refs argument to Cv_bv.

Starting to restrict the domain of the location-to-location map to the
locations actually in the store throughout, in preparation for the
required invariant about not touching those RefPtrs...