make clear a case in cce_aux can't happen, and think about good_code_env
Resigned now to the option of just saying that the output of the
compiler appears, exactly, at the specified label. Any more semantic
characterisation of a good code environment will run into infinite
regress problems, because to prove that the result of compile_code_env
is good will require assuming that the whole code env compile_code_env
produces is good.

Next step: want the code environment in the semantics (the first
argument to Cevaluate) to store the desired closure environment and the
names (and number) of the arguments. Currently it only stores the body
(and the semantics assumes that body is good for whatever closure is the
value of the operand expression). (The implementation, of course,
already keeps track of the env and argument count.)