WIP: attempt at extending semantic code env
Make it store all the information that should be in the closure
associated with the label. There is also an implementation code env, but
this currently still only stores the body (and some other information is
in ecs and env_azs). Also, start thoughts on how to use this extra info
to define good_code_env.

Trying to get the proofs to rebuild, this doesn't seem right. Probably
better to have the extra information in an extra argument to Cevaluate
rather than clobbering the stuff that depends on the way the code
environment currently works. Also might want to make it match env_azs
more closely.