augment definition of good_code_env
impose more conditions on the compiler state when the closure body is
compiled.

seems to require knowing the domain of the closure's environment (to
ensure all binders use distinct names from anything in it), so will need
to add that argument to the d argument of Cevaluate.

also seems to require knowing an upper bound on the compiler's
next_label (to ensure labels generated during closure compilation are
distinct from labels generated in compilation), so add this as an
argument to good_code_env.