rework compile_val statement slightly
put good_cls in the hypotheses and conclusion, and,
require binders to be distinct only from the cs.env (rather than the
whole semantic env). it turns out that this is enough for the CLet case
to go through, and should make it easier to prove the distinctness
condition in the CCall case.