Revert "weaken definition of syneq"
syneq was correct: the closure environment does need to contain the
variables from all the closure bundle bodies, not just the marked one,
because it may want to call the other ones...

The problem is elsewhere, with things like d and good_cls(?) only
looking at the free variables in the marked body.

This reverts commit a94ee69f7682ba429651864b5d6d9c1f5beca092 and
children.