resume work on compile_val
Cenv_bs_bind_fv works.
Now tweaking definition of good_code_env: need to constrain the
next_label of the compiler state when the closure bodies are compiled
(it will be greater than all labels in previous code).
Also want to say that decl will always be NONE when compiling a closure
body... not sure that is true: will enforce it and run the tests...