attempt to prove rest of CIf; found a bug
this commit includes a test expression that fails because of a let
expression in an unused branch of an if expression in tail position: we
need to reset the tail position information before compiling the other
branch.

also the statement of compile_val for the tail code is slightly wrong
since the return code might not come immediately after the code: there
might be a jump (e.g., over the other branch of an if) to the return
code. i think if expressions are the only ones that do that though.