WIP: new approach to how let expressions interact with tail calls
if the input state is TCTail, then you must promise to discard let-bound
variables because the enclosing let is not going to. you can do it
either by making a call (and setting things up correctly before the
call), or by popping them as the last thing you do.

thus the resulting state of a call to compile is always either TCNonTail
or TCTail az env0 0. (Maybe worth another constructor TCEndLet az which
just retains the argument count? or with an option type in TCTail)