rewrite compiler so all exps in tail position return
and function bodies thus don't have to explicitly return.

also, getting clearer about which items in the compiler state are
input-only, i.e. their value in a returned state shouldn't be
informative (or should be the same as in the input state). really only
the compiled code and the next_label counter are for output.

fails the tests currently, but likely just made some minor mistakes.