start work on new good_ecs
Now saying that it is exactly what calculate_ecs produces, namely the
result of ITSET bind_fv. Updated compile_closures_thm to state its
requirements without any mention of good_ecs (and good_ec is now gone
anyway): the hope is to be able to meet these requirements with a
good_ecs.