rework compilation of closures and their environments
Renumber the free variables in the body to match the shape of the
closure environment (as opposed to the enclosing environment): in
particular, remove the "gaps" between free variables for the ones that
aren't captured by the closure.

Compiler tests up to e23 working. e23 involves pattern matching with
multi-variable patterns; I suspect I'm putting the variables in the
wrong order somewhere.