rework declarations, and bug fixes in pattern match compilation
Shadowed declarations are now overwritten on the stack (as before
deBruijn, although the order in which variables appear on the stack
after a Dlet is reversed).

Had a lot of trouble figuring out the deBruijn indices for pattern match
compilation. The tests indicate it is probably right now.

Something is still wrong (tests failing) with recursive closure
environments when there's 3 closures in the bundle - probably off-by-one
or reversed somewhere.