start work on getting intLangTheory to build again
need to think a lot more about how to deal with syneq. at the moment it
may not be easy to prove that it is reflective, and while I could add a
rule forcing it to be reflective, this does not bode well. also, it's
currently not using a proper code environment.