rethinking definition of syneq
the hope is to relate the values/expressions in the intermediate
language that the compiler actually transforms between, to show that
this relation is an equivalence, and that related things evaluate to
related results.

this will be more useful when the compiler does
simplifications/optimizations. at the moment the compiler doesn't do any
interesting transformations on the intermediate language, but it does
change the bodies of closures so the free variables match their
implementation environments, and it also transforms unlabelled
expressions to labelled expressions with a code environment. these
transformations should also be captured by the relation.