Revert "make syneq delete a label from the code env after use"
There's no point making the relation more restrictive for the sake of
reflexivity: there are still bad inputs out there that won't reflect.

So, now I am resigned to thinking that syneq is not reflexive, except
when restricted to "well-formed" inputs. But I don't want to define
"well-formed" since it will involve saying there are no cycles in labels
through the code environment, so it's better to do without a reflective
relation (or take its reflexive closure if it proves necessary to have
reflexivity).