current syneq is unworkable, notes on possible new version
the trouble is with letrec we don't know all the values that should go
into the closure environment (we know that some of them are supposed to
be equivalent), but we need to extend the environment with these
closures to compare the body expressions. the idea for getting around
this is not to use environments at all but just keep track of pairs of
variables that are supposed to be equivalent... still not sure how to
compare those closures though.