work on Cevaluate_syneq
did closure-creation cases, and started on function call case but it
looks like it won't work. might want to modify syneq so that closures
are only equivalent when all their defs are equivalent (this means the
defs on each side must be the same length, and the one picked out must
be the same).