prove mkshift_thm. it only works when there are no labels.
this result is that e is equivalent to (mkshift f k e) if we assume
variables are shifted by f to equivalent values.

it requires there to be no labels because things in the code environment
have already been compiled in the context of the unshifted expression...
I don't think it's necessary (or easy) to post-hoc shift the code
enviroment.