merge closure_extra_data into closure_data
I think it will be easier to have the implementation provide the
not-strictly-necessary semantic data in closure_data records, as witness
that it is doing the right thing, rather than figuring out what's
supposed to be there in the proofs.

Currently providing 0 for the ez component incorrectly, because the
implementation of label_closures doesn't keep track of the size of the
ceration-site environment. But I am hoping that it will be able to in
the next commit...