have another go trying to prove syneq refl (without the extra rule)
This will never work, though, because it's not true: expressions that go
into an infinite loop via the code environment can't have a finite proof
of equivalence to themselves unless the relation hard-codes syntactic
reflexivity somewhere. The principled approach is probably to define a
well-formedness condition for non-loopy code environments, and assume it
everywhere.