rework syneq to only use 1 code env
I think the idea is that the this code env is always fixed, and contains
all the labels that might be introduced.

Happily, I was able to get label_closures_thm proved, even though it
introduces new stuff into the code env! (The key is that label_closures
only operates on completely unlabelled stuff, and when there are no
labels the code env becomes irrelevant.)