work towards thm about compile_code_env, but stuck
Because I want to induct on the code env as a list, but the notion of a
good code env just says the free labs of the bodies are anywhere in the
domain, rather than only in the previous part. I.e. I'm using a finite
map but maybe I care about the order. This would be equivalent to making
sure code envs are well founded. There may be a way around all this in
the proof, or otherwise maybe it is worthwhile going back and using a
list throughout... that will break a lot of stuff slightly...