Initial work on removing the code env.
The idea is to store the closure data in the expression itself, as
opposed to storing labels in the expression that refer to a code
environment mapping labels to closure data. This way, we avoid all the
problems of the map's domain and the labels actually occurring in the
expression getting out of sync.

Have also removed the nz, ez, and ix components of closure data. nz and
ix are obvious from the context of where the thing appears in the
expression. not sure whether ez is actually necessary... will find
out... at least part of it is available from position in the expression
too...