fix minor issues with last commit
next need to define cce_aux, which now needs to traverse the expression
looking for closure data to compile. alternatively could collate them in
a list first... but the traversal needs to happen somewhere, may as well
be here. or label_closures could do it (as before)...

more radically, we could compile the closures on the fly as opposed to
in a separate phase beforehand...