Write a REPL-step function (repl_parse_step) that does REPL parsing.
In particular, its type is

  : token list -> repl_parse_result

where repl_parse_result is either RPR_INCOMPLETE corresponding to the
situation where the input doesn’t appear to have a suitable top-level
semi-colon, or

    RPR_PROG tds toks’

where tds is a ast_prog option, and where toks’ is the rest of the
input (a suffix of the original input).

I don’t particularly expect this to the final word on how the REPL
will consume its input, but I think all the necessary primitives are
illustrated so that new versions will be easy to write.