rewrite compile_decl again, simplify compiler_state
the rsz and renv fields were redundant: they can both be calculated from
rbvars. (the first was its length and the second a list of numbers down
from that length to 1.)

now compile_decl partitions the variables to be bound into those that
are going to shadow old ones and those that are new. it then has
separate helper functions for organising the stack in either case. the
helper functions are simpler because they know what case they're in.