rework repl_fun to take tops one at a time
This change is primarily motivated by making rollback in the case of an
unhandled exception easier, because we can now commit the changes after
each succesfully evaluated declaration/module. I think it also makes the
intended semantics easier to understand. (The repl semantics is not
updated in this commit though.)