make compile_dec deal properly with the possibility of an exception
following the scheme that if the generated code raises an exception, we
reset the stack and put a 1 on top (above the exception), otherwise we
update the stack with the new bindings and put a 0 on top.

compiler (emitml) tests and repl_fun (eval) tests still work.
compiler proofs broken.