Allow for “validating” parser by adding a flag to mmlParseREPLTop.
This function is called from within parse_def, which doesn’t expose
the flag. I did it this way so that the validation/not decision
wouldn’t contaminate higher-level clients, but changing it does
require a big recompilation. Whoever has the lock on REPL
implementation is welcome to make the flag appear in a higher-level
API.