rework compile_primitives, and pre-evaluate it
forgot that it would add a 0 to the top of the stack that the printer is
meant to take care of, so make sure to pop it.

pre-evaluation helps a little bit, but EVALing the examples still takes
way too long because it keeps going through what is now quite
a long list of instructions in the bc_state.