rework generated code for raise/handle
Raise is now responsible for jumping to the handler, which will be a
code pointer under the exception marker.

Also slightly simplify the handler for compile_dec (it was doing PushPtr
JumpPtr instead of just Jump).

Add two simple tests: raising an unhandled exception, and raising and
handling an exception. The latter fails because the exception integer is
off-by-2, because my handlers don't account for the offset due to
Bind_error and Div_error (which are supposed to always be re-raised
anyway).