WIP: start on new statement of compile_val
to account for exceptions. cheat on all the cases in the proof where it
comes up, for now, and try to get the other parts of the original proof
working with the new statement.

might want to tweak the statement more, though, to make the proof avoid
early case splits.