possible bugfix of compilation of Handle
The issue is how to compile a handle expression in tail position.

If no exception is raised, we need to jump to the return pointer of the
last called function.

If an exception is raised, we need to jump to the handler code, and then
it needs to (eventually) jump to the return pointer. (Or raise another
exception or whatever...)

Previously, in the non-exception case, I was jumping to the return
pointer _after_ jumping over the handler code. But that means in the
exception case, when the handler code finishes (and jumps correctly)
there is some leftover junk code after it implementing the original
expression's return. I don't know if this can show up as bad behaviour,
but it at least doesn't get along with the way the proof works (which
assumes the machine runs to the end of the code for an expression).

Now the original expression's return is before the handler code starts,
which seems much more sensible.