addition to statement possibly required by last CCall case
but even with this, it looks like the case when the called function
raises an exception will require duplication of a lot of proof script
(basically both the tail and non-tail non-exception cases). there should
be another way where we can share the proof up to calling the function,
and only split on the behaviour later. not sure if changing the
statement to do the split later will affect the rest of the proof too
much - maybe it's possible to do it only in the CCall case with some
suff_tac manipulations...