half a proof of evaluate_locs
this is the general theorem that Locs in values are still bound by the store
after evaluation. was going to try to use type soundness to get around
having to prove this, but it looks like it's better to just prove this
because it's needed for intermediate steps when evaluating multiple decs
(for a module).

not using it yet because the definition of closed_context needs to be
changed to also constrain locations in the module environment.