work on compile_top_thm, and look at case where module fails
and we add the module name with an empty environment to the rmenv. but
verifying env_rs after that addition looks hard - and I want to make
sure it's necessary first, so will rid of it in next commit. after all, the
implementation only needs the rmenv to look stuff up that it will
actually find... i.e. empty envs are useless..