remove rmenv |+ (mn,[]) from the compiler
that indeed makes that case go through easily.
however, there is still the case where we have to add a real module
environment, and that is probably going to be harder than I thought. At
least, I probably have to prove that the compiler produces the same code
if you give it an extended module environment... or maybe change
code_env_cd (and whatever calls it) to not treat the menv as fixed but
take it in as a parameter..