tweaks to the things in bootstrapProofs for new compiler
Also spent a lot of time thinking about how to state and prove the
theorem that enables you to make a call to a function that has been
compiled already and get the correct behaviour.

Not much progress, but it seems like we'll at least want to prove this
purely in terms of the intermediate language first, and possibly not
even with a function call, but just with evaluating an expression (the
function body) out of context... once such a theorem is at hand, it
might be easier to lift it up to a source language function call. There
will be a lot of delicate assumptions on that theorem, though.

Proved a simple lemma about evaluating a bunch of declarations starting
from initial states, and ending up with the main invariant holding.
(This is the first, easy, part of what's required for bootstrapping.)