strengthen compile_decs_thm by only requiring handler for exceptions
i.e. if the semantics isn't going to produce an exception, the machine
doesn't have to have the handler installed. however, the theorem still
needs a handler for timeout exceptions even though they shouldn't be
necessary, because compile_val is stated weakly (and i don't want to go
touching it right now..)