start on compile_call_val theorem
This theorem is for relating an intermediate language CCall of a
variable bound to a closure on variables bound to arguments that results
in a value, to an execution of bytecode in an appropiate starting state
to yield a refinement of the value. haven't completely figured out
"appropriate starting state" yet.

Likely going to use the Cenv_bs_bind_fv theorem to say the required
environment for evaluating the body can be created by using the values
inside the bytecode closure environment