Start on user-defined exceptions
Update the semantics for raising and handling user-defined exceptions,
but not for defining new ones.  My changes to the grammar and ptree
conversion are just best guesses that I'm not confident in.  I also
added cheats to terminationScript because changing the AST makes HOL
renumber the size functions, breaking the proofs there.