Write a basic shell-script to do some regression testing.
Edit the build-sequence file to add more directories to the test. Each
directory has Holmake cleanAll called it when visited, so the order of
the directories in the sequence should be accurate (or else you'll end
up rebuilding the same ancestral stuff multiple times).

The script will need to be run in a cron job with output sent
somewhere appropriate.